feat(widget): optimize elapsed time calculation in SupplyStateManager

+++ src/client/java/net/iqaddons/mod/state/SupplyStateManager.java
@@ -118,7 +118,7 @@ public final class SupplyStateManager {
     }
 
     public long getElapsedTimeSeconds() {
-        return (long) (getElapsedTimeMillis() / 1000.0);
+        return getElapsedTimeMillis() / 1000;
     }
 
     @Contract(pure = true)
